requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
datasets==2.13.1
dill==0.3.6
filelock==3.9.0
frozenlist==1.4.0
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
networkx==3.0
numpy==1.24.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pyarrow==12.0.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.28.1
safetensors==0.3.1
six==1.16.0
sympy==1.11.1
timm==0.9.2
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
xxhash==3.2.0
yarl==1.9.2
